Expand cli/workspaces/index, delete lang/workspaces/via-backends
The majority of the content in the proposed via-backends.html page dates back to
a 2018 edit (0a5059df151a1eb3253f2d73db4e42c97cf67cc4) aimed at temporarily
patching an understanding gap that no longer seems to exist (or at least not in
that particular form).

This commit is an effort to retain the most important CLI-specific points from
that content in a more concise form, and put them where they'll be findable to
CLI users.